+++ VAutomationCore/Json/PrefabGuidJsonConverter.cs
@@ -6,7 +6,7 @@ using Stunlock.Core;
 namespace VAuto.Core.Json
 {
     /// <summary>
-    /// Serializes PrefabGUID as an int (GuidHash), and deserializes from:
+    /// Serializes PrefabGUID as a long (GuidHash), and deserializes from:
     /// - number: -123
     /// - string: "-123" or "Prefab_Name"
     /// - object: { "guidHash": -123 } / { "GuidHash": -123 }
@@ -19,7 +19,7 @@ namespace VAuto.Core.Json
             {
                 if (reader.TokenType == JsonTokenType.Number)
                 {
-                    return new PrefabGUID(reader.GetInt32());
+                    return new PrefabGUID((int)reader.GetInt32());
                 }
 
                 if (reader.TokenType == JsonTokenType.String)
@@ -29,7 +29,7 @@ namespace VAuto.Core.Json
                         return default;
 
                     if (int.TryParse(s, out var i))
-                        return new PrefabGUID(i);
+                        return new PrefabGUID((int)i);
 
                     // Allow config to use prefab names.
                     if (PrefabGuidConverter.TryGetGuid(s, out var guid))
@@ -44,7 +44,7 @@ namespace VAuto.Core.Json
                     var root = doc.RootElement;
 
                     if (TryReadGuidHash(root, "guidHash", out var gh) || TryReadGuidHash(root, "GuidHash", out gh))
-                        return new PrefabGUID(gh);
+                        return new PrefabGUID((int)gh);
 
                     // As a last resort, accept { "name": "Prefab_Name" }
                     if (root.TryGetProperty("name", out var nameEl) && nameEl.ValueKind == JsonValueKind.String)
@@ -72,16 +72,16 @@ namespace VAuto.Core.Json
             writer.WriteNumberValue(value.GuidHash);
         }
 
-        private static bool TryReadGuidHash(JsonElement root, string property, out int guidHash)
+        private static bool TryReadGuidHash(JsonElement root, string property, out long guidHash)
         {
             guidHash = default;
             if (!root.TryGetProperty(property, out var el))
                 return false;
 
-            if (el.ValueKind == JsonValueKind.Number && el.TryGetInt32(out guidHash))
+            if (el.ValueKind == JsonValueKind.Number && el.TryGetInt64(out guidHash))
                 return true;
 
-            if (el.ValueKind == JsonValueKind.String && int.TryParse(el.GetString(), out guidHash))
+            if (el.ValueKind == JsonValueKind.String && long.TryParse(el.GetString(), out guidHash))
                 return true;
 
             return false;
